move some chaning.py + updater files  into animation
