+++ manim/animation/changing.py
@@ -8,10 +8,10 @@ from typing import Callable
 
 from colour import Color
 
-from .._config import config
-from ..mobject.types.vectorized_mobject import VGroup, VMobject
-from ..utils.color import BLUE_B, BLUE_D, BLUE_E, GREY_BROWN, WHITE
-from ..utils.rate_functions import smooth
+from manim._config import config
+from manim.mobject.types.vectorized_mobject import VGroup, VMobject
+from manim.utils.color import BLUE_B, BLUE_D, BLUE_E, GREY_BROWN, WHITE
+from manim.utils.rate_functions import smooth
 from manim.mobject.opengl.opengl_compatibility import ConvertToOpenGL
