+++ manim/camera/three_d_camera.py
@@ -17,7 +17,7 @@ from manim.mobject.three_d.three_d_utils import (
     get_3d_vmob_start_corner_unit_normal,
 )
 from ..mobject.types.point_cloud_mobject import Point
-from ..mobject.value_tracker import ValueTracker
+from manim.animation.updaters.value_tracker import ValueTracker
 from ..utils.color import get_shaded_rgb
 from ..utils.family import extract_mobject_family_members
 from ..utils.space_ops import rotation_about_z, rotation_matrix
