+++ manim/mobject/text/numbers.py
@@ -12,7 +12,7 @@ from manim import config
 from manim.constants import *
 from manim.mobject.text.tex_mobject import MathTex, SingleStringMathTex
 from manim.mobject.types.vectorized_mobject import VMobject
-from manim.mobject.value_tracker import ValueTracker
+from manim.animation.updaters.value_tracker import ValueTracker
 from manim.mobject.opengl.opengl_compatibility import ConvertToOpenGL
 
 string_to_mob_map = {}
